@mmedenjak 2019-02-08
@jerrinot is this PR still WIP? Can it be closed until it is finalised?