@mmedenjak 2020-04-06
Since we configured the build environment to lower resource consumption, we can park this PR until we see the effects of the change first.